+++ builder_pkgs/mockito/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Require analyzer 5.12.0, allow analyzer version 6.x;
 * Add example of writing a class to mock function objects.
+* Add support for the `build_extensions` build.yaml option
 
 ## 5.4.2